% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BrestSurgeDays}
\alias{BrestSurgeDays}
\title{Number of days per month in 1846-2007}
\format{
\code{BrestSurgeDays} is a data frame with 162 rows (years 1846 to
2007) and the 12 variables (one for each month of the year). Each value
in the data frame gives the number of days in the month in question.

The row names of \code{BrestSurgeMaxima} are the years \code{1946:2007} and the column
names are the abbreviated names of the months.
}
\usage{
BrestSurgeDays
}
\description{
Number of days in each month relevant to the Brest sea surge heights data
\code{\link{BrestSurgeMaxima}}.
}
\examples{
head(BrestSurgeDays)
}
\seealso{
\itemize{
\item \code{\link{BrestSurgeMaxima}}: Annual maxima surge heights at Brest, France.
\item \code{\link{BrestSurgeMissing}}: numbers of missing values in each month.
}
}
\keyword{datasets}
