% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LaPlagne}
\alias{LaPlagne}
\title{Winter Snow Depth and Precipitation at La Plagne, France}
\format{
\code{LaPlagne} is a data frame with 4386 rows and the 5 variables:
\itemize{
\item \code{rr24}: precipitation in millimetres in the previous 24 hours.
\item \code{ht_neige}: total snow depth in metres.
\item \code{date}: with class \code{"Date"} in the format \code{YYYY-MM-DD}.
\item \code{time}: time at which the observations are recorded, in the format
\code{hh:mm:ss}.
\item \code{winter}: the year in which the winter starts.
}

18.5\% of the precipitation values and 11.8\% of the snow depth data are
missing.
}
\source{
METEO FRANCE’s
\href{https://meteo.data.gouv.fr/datasets/68aeced879316c333d564f95}{snow-weather network archive}.
}
\usage{
data(LaPlagne)
}
\description{
Daily snow depth and precipitation at La Plagne (France) for the 29 winters
(December to April inclusive) starting in 1996 to 2024
}
\examples{
# Data strcture
head(LaPlagne)
tail(LaPlagne)

# Percentage missing values
mean(is.na(LaPlagne$rr24))
mean(is.na(LaPlagne$ht_neige))

# Cross-tabulation of precipitation and snow missingness
table(is.na(LaPlagne$rr24), is.na(LaPlagne$ht_neige))

# Time series plots
plot(LaPlagne$date, LaPlagne$rr24, pch = 16,
     xlab = "year", ylab = "24 hour precipitation (mm)")
plot(LaPlagne$date, LaPlagne$ht_neige, pch = 16,
     xlab = "year", ylab = "snow depth (m)")
}
\seealso{
\code{\link{LaPlagnePrecipMissings}}, \code{\link{LaPlagneSnowMissings}},
\code{\link{LaPlagnePrecipMaxima}}, \code{\link{LaPlagnePrecipMaximaSeasonal}},
\code{\link{LaPlagneSnowMaxima}}, \code{\link{LaPlagneSnowMaximaSeasonal}}.
}
\keyword{datasets}
