% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LaPlagnePrecipMaxima}
\alias{LaPlagnePrecipMaxima}
\title{Winter Precipitation Maxima and Missing Information at La Plagne, France}
\format{
\code{LaPlagnePrecipMaxima} is a list inheriting from class
\code{"block_maxima"}. See \code{\link{block_maxima}} and \code{\link{plot.block_maxima}}.
}
\usage{
data(LaPlagnePrecipMaxima)
}
\description{
These data are the result of a call
\code{block_maxima(data = LaPlagne$rr24, block = LaPlagne$winter, pseudo = TRUE, seasonal = FALSE)}
to \code{\link{block_maxima}}.
}
\seealso{
\code{\link{LaPlagne}}.
}
\keyword{datasets}
