% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LaPlagnePrecipMissings}
\alias{LaPlagnePrecipMissings}
\title{Winter Precipitation Missingness at La Plagne, France}
\format{
\code{LaPlagnePrecipMissings} is a data frame with 813 rows and the 2
variables:
\itemize{
\item \code{winter}: the winter in which a missing precipitation value occurs.
\item \verb{day of winter}: the day of the winter on which the missing value occurs.
}
}
\usage{
data(LaPlagnePrecipMissings)
}
\description{
Provides the days on which the 24 hour precipitation variable is missing,
that is, the winter in which is occurs and the day of this winter, where
day 1 is 1 Decemeber, day 2 is 2 December etc.
}
\examples{
# Pattern of missingness over and within winters
plot(LaPlagnePrecipMissings, pch = "_", main = "24 hour precipitation")
}
\seealso{
\code{\link{LaPlagne}}.
}
\keyword{datasets}
