% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LaPlagneSnowMaxima}
\alias{LaPlagneSnowMaxima}
\title{Winter Snow Depth Maxima and Missing Information at La Plagne, France}
\format{
\code{LaPlagneSnowMaxima} is a list. See \code{\link{block_maxima}}.
}
\usage{
data(LaPlagneSnowMaxima)
}
\description{
These data are the result of a call
\code{block_maxima(data = LaPlagne$ht_neige, block = LaPlagne$winter, pseudo = TRUE, seasonal = FALSE)}
to \code{\link{block_maxima}}.
}
\seealso{
\code{\link{LaPlagne}}.
}
\keyword{datasets}
