% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LaPlagneSnowMaximaSeasonal}
\alias{LaPlagneSnowMaximaSeasonal}
\title{Winter Snow Depth Maxima and Seasonal Missing Information at La Plagne, France}
\format{
\code{LaPlagneSnowMaximaSeasonal} is a list. See \code{\link{block_maxima}}.
}
\usage{
data(LaPlagneSnowMaximaSeasonal)
}
\description{
These data are the result of a call
\code{block_maxima(data = LaPlagne$ht_neige, block = LaPlagne$winter, pseudo = TRUE, seasonal = TRUE)}
to \code{\link[=block_maxima]{block_maxima()}}.
}
\seealso{
\code{\link{LaPlagne}}.
}
\keyword{datasets}
