% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_block_maxima.R
\name{plot.block_maxima}
\alias{plot.block_maxima}
\title{Plot method for objects inheriting from class \code{"block_maxima"}}
\usage{
\method{plot}{block_maxima}(x, which = 1, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{"block_maxima"}.}

\item{which}{If \code{which = 1} then the sliding block maxima are plotted
against block number. If \code{which = 2} then the sliding block maxima are
plotted against the proportion of non-missing raw values.}

\item{...}{Further arguments to \code{\link[graphics:plot.default]{plot}}.}
}
\value{
Nothing is returned.
}
\description{
Plot method for objects inheriting from \code{"block_maxima"} returned from
\code{\link{block_maxima}}, \code{\link{block_maxima_ts}} or \code{\link{sliding_block_maxima_ts}}.
}
\details{
When \code{which = 1} we obtain a time series plot in which there are
periods where the value does not change. When \code{which = 2} we expect to
see that the sliding block maximum tends to be smaller for blocks with a
larger proportionof missing values.
}
\examples{
### Plymouth Ozone Data

## Sliding maxima
# Time series plots of sliding block maxima
plot(PlymouthOzoneSlidingMaxima)
# Plot maxima against the proportion of non-missing daily values
plot(PlymouthOzoneSlidingMaxima, which = 2)

## Disjoint maxima
bm <- block_maxima(PlymouthOzone$Ozone, block = PlymouthOzone$Year)
# Time series plots of block maxima
plot(bm)
# Plot maxima against the proportion of non-missing daily values
plot(bm, which = 2)
}
