% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_block_maxima.R
\name{plot.sliding_block_maxima}
\alias{plot.sliding_block_maxima}
\title{Plot method for objects of class \code{"sliding_block_maxima"}}
\usage{
\method{plot}{sliding_block_maxima}(x, which = 1, ...)
}
\arguments{
\item{x}{An object inheriting from class \code{"sliding_block_maxima"}, an
object from a call to \code{\link{sliding_block_maxima_ts}}.}

\item{which}{If \code{which = 1} then the sliding block maxima are plotted
against block number. If \code{which = 2} then the sliding block maxima are
plotted against the proportion of non-missing raw values.}

\item{...}{Further arguments to \code{\link[graphics:plot.default]{plot}}.}
}
\value{
Nothing is returned.
}
\description{
Plot method for objects of class \code{"sliding_block_maxima"} returned from
\code{\link{sliding_block_maxima_ts}}, for example \code{\link{PlymouthOzoneSlidingMaxima}}.
}
\details{
When \code{which = 1} we obtain a time series plot in which there are
periods where the value does not change. When \code{which = 2} we expect to
see that the sliding block maximum tends to be smaller for blocks with a
larger proportionof missing values.
}
\examples{
# Time series plots of sliding block maxima
plot(PlymouthOzoneSlidingMaxima)

# Plot maxima against the proportion of non-missing daily values
plot(PlymouthOzoneSlidingMaxima, which = 2)
}
