% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PlymouthOzoneSlidingMaxima}
\alias{PlymouthOzoneSlidingMaxima}
\title{Sliding Maxima of Ozone Levels at Plymouth, UK}
\format{
\code{PlymouthOzoneSlidingMaxima} is an object returned from
\code{\link{sliding_block_maxima_ts}}, a list inheriting from class
\code{sliding_maxima_ts}. It was created using
\code{sliding_block_maxima_ts(PlymouthOzone$Ozone, block_length = 365)}.
This includes components
\itemize{
\item \code{full_maxima}: a numeric vector of 130 full-block maxima, all equal to 143
\eqn{\mu}g/m\eqn{^3}.
\item \code{partial_maxima}: a numeric vector of 9368 partial-block maxima.
\item \code{maxima}: a numeric vector of all maxima.
}

See \code{\link{sliding_block_maxima_ts}} for a description of the other components.
}
\source{
The Department for Environment Food and Rural Affair (DEFRA).
The Plymouth Centre monitoring site at
the \href{https://uk-air.defra.gov.uk/}{UK-AIR} database
\href{https://uk-air.defra.gov.uk/data/data_selector}{Data Selector}.
}
\usage{
PlymouthOzoneSlidingMaxima
}
\description{
Maxima of sliding (overlapping) block of length 365 days of daily maximum
ozone levels at Plymouth in Devon (UK) for the years 1998-2024 inclusive.
}
\examples{
# Time series plots of sliding block maxima
plot(PlymouthOzoneSlidingMaxima)

# Plot maxima against the proportion of non-missing daily values
plot(PlymouthOzoneSlidingMaxima, which = 2)
}
\seealso{
\code{\link{PlymouthOzone}} for the raw time series.
}
\keyword{datasets}
