% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_block_maxima_ts.R
\name{sliding_block_maxima_ts}
\alias{sliding_block_maxima_ts}
\title{Sliding Block Maxima for a Time Series}
\usage{
sliding_block_maxima_ts(data, block_length)
}
\arguments{
\item{data}{A numeric vector containing a time series of raw data.}

\item{block_length}{A numeric scalar. Used calculate the maxima of sliding
blocks of \code{block_length} contiguous values in the vector \code{data}.}
}
\value{
A list, with class
\code{c("list", "block_maxima", "sliding", "evmissing")}, containing the
following components:
\itemize{
\item \code{maxima}: the block maxima.
\item \code{notNA}: the numbers of non-missing observations in each block.
\item \code{n}: the maximal block length, that is, the largest number of values that
could have been observed in each block.
\item \code{whereNA}: a named list containing, for each block, the positions of any
missing values in the block. For example, if (only) the first and fifth
observations in block 3 are missing then the third component (named
\code{block3}) of \code{whereNA} is \code{c(1, 5)}. If a block has no missing values
then its component in \code{whereNA} is \code{integer(0)}.
\item \code{pseudo_maxima}: a numeric matrix containing (pseudo) block maxima
created by applying the missing value patterns from incomplete blocks to
all full blocks, that is, blocks without any missing values. Each column
contains the pseudo-maxima resulting from a particular incomplete block.
The columns are labelled by the number of the incomplete block and the
columns by the number of the full block. If an incomplete block contains
all missing values then its entry in \code{pseudo_maxima} is \code{NA}. If there
are no full blocks or no incomplete blocks then \code{pseudo_maxima} is \code{NA}.
\item \code{full_maxima}: a numeric vector of maxima from full blocks.
\item \code{partial_maxima}: a numeric vector of maxima from partial blocks.
}

If a block contains only missing values then its value of \code{maxima} is \code{NA},
its value of \code{notNA} is \code{0} and \code{whereNA} contains the positions of all the
observations in the block.

If \code{block} is supplied then these vectors are named using the values in
\code{block}. Otherwise, these vectors do not have names.
}
\description{
Extracts sliding block maxima and missing value information for each block.
Works like \code{\link{block_maxima_ts}} but returns information for sliding
(overlapping) blocks rather than disjoint blocks and defines blocks using
only block length, that is, there is no \code{block} argument.
}
\details{
The maxima are calculated for \strong{all} blocks of length
\code{block_length} present in \code{data}, starting with the first block
\code{data[1:block_length]} and sliding the block repeatedly by one observation
until reaching the final block
\code{data[(length(data) - block_length + 1):length(data)]}.
}
\examples{
## Simulate example data
set.seed(7032025)
data <- rexp(15)

# Create some missing values
data[c(5, 7:8)] <- NA
# 5 blocks (columns), each with 3 observations
matrix(data, ncol = 5)

block_length <- 3
x <- sliding_block_maxima_ts(data, block_length = block_length)
}
