% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_maxima.R
\name{block_maxima}
\alias{block_maxima}
\title{Block Maxima and Missing Information}
\usage{
block_maxima(
  data,
  block_length,
  block,
  pseudo = FALSE,
  full = FALSE,
  sliding = FALSE,
  seasonal = sliding
)
}
\arguments{
\item{data}{A numeric vector containing a time series of raw data.}

\item{block_length}{A numeric scalar. Used calculate the maxima of
disjoint blocks of \code{block_length} contiguous values in the vector \code{data}.
If \code{length(data)} is not an integer multiple of \code{block_length} then
the values at the end of \code{data} that do not constitute a complete block
of length \code{block_length} are discarded, without warning.}

\item{block}{A numeric vector with the same length as \code{data}. The value of
\code{block[i]} indicates the block into which \code{data[i]} falls. The block
lengths implied by \code{block} may differ by at most 1. For example,
\code{block[i]} could give the block in which observation \code{i} was observed,
with block lengths of 366 for leap years and 365 for other years.}

\item{pseudo}{A logical scalar. If \code{pseudo = TRUE} then pseudo-maxima are
calculated as the block maxima obtained by applying the missing value
patterns from partial disjoint blocks to all suitable other blocks, which
we call \strong{donor} blocks. See \code{full}.}

\item{full}{If \code{full = FALSE} then a donor block for a given partial
disjoint block is any block that has non-missing values in all the
comparable positions where the partial block has non-missing values.
The comparable positions depend on whether we respect the seasonality that
may be exhibited within a block. See \code{seasonal}.
If \code{full = TRUE} then only full blocks are used as donor blocks:
if \code{sliding = FALSE} then \code{data} must contain at least one full disjoint
block and if \code{sliding = TRUE} then \code{data} must contain at least one full
sliding block.}

\item{sliding}{A logical scalar. Only relevant if \code{pseudo = TRUE}. If
\code{sliding = TRUE} then pseudo-maxima are calculated for \strong{all} donor blocks
of the relevant length, rather than only a set of disjoint donor blocks.}

\item{seasonal}{A logical scalar. Only relevant if \code{pseudo = TRUE} and
\code{sliding = TRUE}. If \code{seasonal = TRUE} then the way in which the
pseudo-maxima are calculated respects the seasonality that may be
exhibited over the duration of a block. If, for example, a block covers a
single year, then the missing values applied to a donor block occur at the
same time of year as in the originating partial block. If
\code{seasonal = FALSE} then the missing values applied have the same positions
in the partial and donor blocks with respect to the start of each block.}
}
\value{
A list, with class
\code{c("list", "block_maxima", "disjoint", "evmissing")},
if \code{sliding = FALSE}, or
\code{c("list", "block_maxima", "sliding", "evmissing")},
if \code{sliding = TRUE},
containing the following numeric vectors:
\itemize{
\item \code{maxima}: the block maxima.
\item \code{notNA}: the numbers of non-missing observations in each block.
\item \code{n}: the maximal block length, that is, the largest number of values that
could have been observed in each block.
}

If \code{block} is supplied then these vectors are named using the values in
\code{block}. Otherwise, they do not have names.

If \code{pseudo = TRUE} then the returned list also contains the following:
\itemize{
\item \code{whereNA}: a named list containing, for each block, the positions of any
missing values in the block. For example, if (only) the first and fifth
observations in block 3 are missing then the third component (named
\code{block3}) of \code{whereNA} is \code{c(1, 5)}. If a block has no missing values
then its component in \code{whereNA} is \code{integer(0)}.
\item \code{pseudo_maxima}: a numeric matrix containing block maxima created by
applying the missing value patterns from partial blocks to all full
blocks. Each column contains the pseudo-maxima resulting from a
particular partial block. The columns are labelled by the number of the
partial block and the rows by the number of the full (disjoint or
sliding) block. If a partial block contains all missing values then its
entry in \code{pseudo_maxima} is \code{NA}. If there are no full blocks or no
partial blocks then \code{pseudo_maxima} is \code{NA}.
\item \code{full_maxima}: a numeric vector of maxima from full blocks.
\item \code{partial_maxima}: a numeric vector of maxima from partial blocks.
}

The input arguments \code{pseudo}, \code{full}, \code{sliding} and \code{seasonal} are also
included.

If a block contains only missing values then its value of \code{maxima} is \code{NA}
and its value of \code{notNA} is \code{0}.
}
\description{
Extracts disjoint block maxima from a time series of raw data. Can also
provide information about the effect of missing value patterns on block
maxima via pseudo-maxima created by applying blockwise missing value
patterns in partially-observed (partial) disjoint blocks to fully-observed
(full) blocks or to full blocks and other suitable partially-observed
blocks.
}
\details{
Exactly one of the arguments \code{block_length} or \code{block} must be
supplied.

If \code{block_length} is supplied and \code{sliding = TRUE} then the pseudo-maxima
are calculated for \strong{all} blocks of length \code{block_length} present in
\code{data}, starting with the first block \code{data[1:block_length]} and sliding
the block repeatedly by one observation until reaching the final block
\code{data[(length(data) - block_length + 1):length(data)]}.

\strong{Also explain for \code{block}}
\subsection{Calculation of pseudo-maxima}{

Consider example data \code{c(1, 2, NA, 3, 6, 5, 8, 7, 4, 3, NA, NA)} and a
block length of 4. There are 3 disjoint blocks, with block maxima
\code{c(3, 8, 4)}. Blocks 1 and 3 are partially-observed. We consider the 6
valid combinations of \code{full}, \code{sliding} and \code{seasonal}.
\subsection{Full donor blocks only}{
\itemize{
\item \code{sliding = FALSE}. Disjoint block 2, \code{c(6, 5, 8, 7)}, is the only donor
block. Applying the missing value patterns from disjoint blocks 1 and 3
to block 2 leads to respective pseudo-maxima \code{c(7, 6)}.
\item \code{sliding = TRUE} and \code{seasonal = FALSE}. There are 4 full sliding donor
blocks. For example, sliding block 4, \code{c(3, 6, 5, 8)}, leads to
pseudo-maxima \code{c(8, 6)} for disjoint blocks 1 and 3. The full sets of
pseudo-maxima are \code{c(8, 7, 8, 8)} for block 1 and \code{c(6, 6, 8, 8)} for
block 3.
\item \code{sliding = TRUE} and \code{seasonal = TRUE}. The vector of seasons is
\verb{c(1:4, 1;4, 1:4)}. Sliding block 4, \code{c(3, 6, 5, 8)} has season vector
\code{c(4, 1, 2, 3)} leads to pseudo-maxima \code{c(6, 6)} for disjoint blocks 1
and 3 because the \code{8} becomes \code{NA} for block 1 and both \code{3} and \code{8}
become \code{NA} for block 3. The full sets of pseudo-maxima are
\code{c(6, 6, 6, 7)} for block 1 and \code{c(6, 6, 5, 4)} for block 3.
}
}

\subsection{All suitable donor blocks}{
\itemize{
\item \code{sliding = FALSE}. Disjoint block 1, \code{c(1, 2, NA, 3)}, can donate to
disjoint block 3, \code{c(4, 3, NA, NA)}, leading to a pseudo-maximum of
\code{2}. The full sets of pseudo-maxima are \code{c(7)} for block 1 and \code{c(2, 6)}
for block 3.
\item \code{sliding = TRUE} and \code{seasonal = FALSE}. In addition to the full sliding
blocks, sliding blocks 1 \code{c(1, 2, NA, 3)} and 8, \code{c(7, 4, 3, NA)}, can
donate to disjoint block 3, \code{c(4, 3, NA, NA)}. The full sets of
pseudo-maxima are \code{c(8, 7, 8, 7, 8)} for block 1 and
\code{c(2, 6, 6, 8, 8, 7)} for block 3.
\item \code{sliding = TRUE} and \code{seasonal = TRUE}. In addition to the full sliding
blocks, sliding block 1 can donate to disjoint block 3 and sliding
blocks 2, 3 and 8 donate to both disjoint blocks 1 and 3. The full sets
of pseudo-maxima are \code{c(6, 6, 6, 7, 7, 7, 7)} for block 1 and
\code{c(2, 6, 6, 6, 6, 5, 4, 4)} for block 3.
}

See \code{\link{gev_ts}} for an explanation of how the pseudo-maxima are used.
}

}
}
\examples{
## Simulate example data
set.seed(7032025)
data <- rexp(15)

# Create some missing values
data[c(5, 7:8)] <- NA
# 5 blocks (columns), each with 3 observations
matrix(data, ncol = 5)

# Supplying block_length, disjoint maxima
block_length <- 3
block_maxima(data, block_length = block_length)

# Supplying block_length, sliding maxima
res1 <- block_maxima(data, block_length = block_length, pseudo = TRUE)
res2 <- block_maxima(data, block_length = block_length, pseudo = TRUE,
                     sliding = TRUE)
# res1 and res2 have the same missing positions, full and partial maxima
res1$whereNA
res1$full_maxima
res1$partial_maxima
# res2 has more pseudo maxima than res1
res1$pseudo_maxima
res2$pseudo_maxima

# Supplying block
block <- rep(1:5, each = 3)
block_maxima(data, block = block)

## Data with a partially-observed block
data <- c(data, 1:2)

# Supplying block_length (the extra 2 observations are ignored)
block_length <- 3
block_maxima(data, block_length = block_length)
# Supplying block (with an extra group indicator)
block <- c(block, 7, 7)
block_maxima(data, block = block)
}
\seealso{
Plot method \code{\link{plot.block_maxima}}.
}
