% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_maxima.R
\name{block_maxima}
\alias{block_maxima}
\title{Block Maxima and missing information}
\usage{
block_maxima(data, block_length, block, sliding = FALSE, pseudo = FALSE)
}
\arguments{
\item{data}{A numeric vector containing a time series of raw data.}

\item{block_length}{A numeric scalar. Used calculate the maxima of
blocks of \code{block_length} contiguous values in the vector \code{data}.
If \code{length(data)} is not an integer multiple of \code{block_length} then
the values at the end of \code{data} that do not constitute a complete block
of length \code{block_length} are discarded, without warning.}

\item{block}{A numeric vector with the same length as \code{data}. The value of
\code{block[i]} indicates the block into which \code{data[i]} falls. For example,
\code{block} could provide the year in which observation \code{i} was observed.}

\item{sliding}{A logical scalar. This is only allowed if \code{block_length} is
supplied. If \code{sliding = TRUE} then maxima are calculated for \strong{all}
blocks of length \code{block_length}. See \strong{Details}.}

\item{pseudo}{A logical scalar. If \code{pseudo = TRUE} then}
}
\value{
A list, with class \code{c("list", "block_maxima", "evmissing")},
containing the following numeric vectors:
\itemize{
\item \code{maxima}: the block maxima.
\item \code{notNA}: the numbers of non-missing observations in each block.
\item \code{n}: the maximal block length, that is, the largest number of values that
could have been observed in each block.
}

If \code{pseudo = TRUE} then the returned list also contains the following:
\itemize{
\item \code{whereNA}: a named list containing, for each block, the positions of any
missing values in the block. For example, if (only) the first and fifth
observations in block 3 are missing then the third component (named
\code{block3}) of \code{whereNA} is \code{c(1, 5)}. If a block has no missing values
then its component in \code{whereNA} is \code{integer(0)}.
\item \code{pseudo_maxima}: a numeric matrix containing (pseudo) block maxima
created by applying the missing value patterns from incomplete blocks to
all full blocks, that is, blocks without any missing values. Each column
contains the pseudo-maxima resulting from a particular incomplete block.
The columns are labelled by the number of the incomplete block and the
columns by the number of the full block. If an incomplete block contains
all missing values then its entry in \code{pseudo_maxima} is \code{NA}. If there
are no full blocks or no incomplete blocks then \code{pseudo_maxima} is \code{NA}.
\item \code{full_maxima}: a numeric vector of maxima from full blocks.
\item \code{partial_maxima}: a numeric vector of maxima from partial blocks.
}

The input arguments \code{sliding} and \code{pseudo} are also included.

If a block contains only missing values then its value of \code{maxima} is \code{NA}
and its value of \code{notNA} is \code{0}.

If \code{block} is supplied then these vectors are named using the values in
\code{block}. Otherwise, these vectors do not have names.
}
\description{
Extracts disjoint or sliding block maxima and information relating to the
pattern of missing values. The (pseudo-)maxima created by applying blockwise
missing value patterns in incomplete block to full blocks can be calculated.
To be useful, the input data, \code{data} must contain at least one full block.
}
\details{
Exactly one of the arguments \code{block_length} or \code{block} must be
supplied.

If \code{block_length} is supplied and \code{sliding = TRUE} then the maxima are
calculated for \strong{all} blocks of length \code{block_length} present in \code{data},
starting with the first block \code{data[1:block_length]} and sliding the block
repeatedly by one observation until reaching the final block
\code{data[(length(data) - block_length + 1):length(data)]}.
}
\examples{
## Simulate example data
set.seed(7032025)
data <- rexp(15)

# Create some missing values
data[c(5, 7:8)] <- NA
# 5 blocks (columns), each with 3 observations
matrix(data, ncol = 5)

# Supplying block_length, disjoint maxima
block_length <- 3
block_maxima(data, block_length = block_length)
# Supplying block_length, sliding maxima
block_maxima(data, block_length = block_length, sliding = TRUE)

# Supplying block
block <- rep(1:5, each = 3)
block_maxima(data, block = block)

## Data with an incomplete block
data <- c(data, 1:2)

# Supplying block_length (the extra 2 observations are ignored)
block_length <- 3
block_maxima(data, block_length = block_length)
# Supplying block (with an extra group indicator)
block <- c(block, 7, 7)
block_maxima(data, block = block)

}
\seealso{
Plot method \code{\link{plot.block_maxima}}.
}
