% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_maxima_ts.R
\name{block_maxima_ts}
\alias{block_maxima_ts}
\title{Block Maxima for a Time Series}
\usage{
block_maxima_ts(data, block_length, block)
}
\arguments{
\item{data}{A numeric vector containing a time series of raw data.}

\item{block_length}{A numeric scalar. Used calculate the maxima of disjoint
blocks of \code{block_length} contiguous values in the vector \code{data}.
If \code{length(data)} is not an integer multiple of \code{block_length} then
the values at the end of \code{data} that do not constitute a complete block
of length \code{block_length} are discarded, without warning.}

\item{block}{A numeric vector with the same length as \code{data}. The value of
\code{block[i]} indicates the block into which \code{data[i]} falls. For example,
\code{block} could provide the year in which observation \code{i} was observed.}
}
\value{
A list, with class \code{c("list", "block_maxima", "evmissing")},
containing the following components:
\itemize{
\item \code{maxima}: the block maxima.
\item \code{notNA}: the numbers of non-missing observations in each block.
\item \code{n}: the maximal block length, that is, the largest number of values that
could have been observed in each block.
\item \code{whereNA}: a named list containing, for each block, the positions of any
missing values in the block. For example, if (only) the first and fifth
observations in block 3 are missing then the third component (named
\code{block3}) of \code{whereNA} is \code{c(1, 5)}. If a block has no missing values
then its component in \code{whereNA} is \code{integer(0)}.
\item \code{pseudo_maxima}: a numeric matrix containing (pseudo) block maxima
created by applying the missing value patterns from incomplete blocks to
all full blocks, that is, blocks without any missing values. Each column
contains the pseudo-maxima resulting from a particular incomplete block.
The columns are labelled by the number of the incomplete block and the
columns by the number of the full block. If an incomplete block contains
all missing values then its entry in \code{pseudo_maxima} is \code{NA}. If there
are no full blocks or no incomplete blocks then \code{pseudo_maxima} is \code{NA}.
\item \code{full_maxima}: a numeric vector of maxima from full blocks.
\item \code{partial_maxima}: a numeric vector of maxima from partial blocks.
}

If a block contains only missing values then its value of \code{maxima} is \code{NA},
its value of \code{notNA} is \code{0} and \code{whereNA} contains the positions of all the
observations in the block.

If \code{block} is supplied then these vectors are named using the values in
\code{block}. Otherwise, these vectors do not have names.
}
\description{
Extracts block maxima and missing value information for each block.
Works like \code{\link{block_maxima}} but returns extra components, including:
\code{whereNA}, the positions of the missing values within each block, and
\code{pseudo_maxima}, the maxima created by applying blockwise missing value
patterns in incomplete blocks to full blocks, that is, blocks without any
missing values. To be useful, the input data, \code{data}, must contain at least
one full block.
}
\details{
Exactly one of the arguments \code{block_length} or \code{block} must be
supplied. If the block sizes implied by \code{block} are unequal then an
incomplete block and a full block may have different lengths. If this
occurs when \code{pseudo_maxima} are calculated, then the longer block is
trimmed, by discarding trailing values, so that the lengths match.
}
\examples{
## Simulate example data
set.seed(7032025)
data <- rexp(15)

# Create some missing values
data[c(5, 7:8)] <- NA
# 5 blocks (columns), each with 3 observations
matrix(data, ncol = 5)
# Supplying block_length
block_length <- 3
block_maxima_ts(data, block_length = block_length)
# Supplying block
block <- rep(1:5, each = 3)
block_maxima_ts(data, block = block)

## Data with an incomplete block
data <- c(data, 1:2)

# Supplying block_length (the extra 2 observations are ignored)
block_length <- 3
block_maxima_ts(data, block_length = block_length)
# Supplying block (with an extra group indicator)
block <- c(block, 7, 7)
block_maxima_ts(data, block = block)
}
