% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evmissing-internal.R
\name{evmissing-internal}
\alias{evmissing-internal}
\alias{negated_gev_loglik}
\alias{weighted_negated_gev_loglik}
\alias{discard_maxima}
\alias{negated_gev_loglik_ret_levs}
\alias{weighted_negated_gev_loglik_ret_levs}
\alias{find_block_i_disjoint}
\alias{find_maxima_notNA_disjoint_pseudo}
\alias{find_maxima_notNA_disjoint}
\alias{x_donates_to_y}
\alias{x_donates_to_y_seasonal}
\alias{find_pseudo_maxima_block_length}
\alias{x_donates_to_y_block}
\alias{x_donates_to_y_seasonal_block}
\alias{adjust_seasonal_positions}
\alias{find_pseudo_maxima_block}
\alias{maxima_block_length}
\alias{maxima_block}
\alias{rhat}
\alias{negated_gev_loglik_ts}
\alias{faster_profile_ci}
\alias{profile_ci}
\alias{box_cox}
\alias{box_cox_vec}
\alias{box_cox_deriv}
\alias{gev_init}
\alias{gev_profile_init}
\alias{call_gev_profile_init}
\alias{return_level_profile_init}
\alias{gev_pp}
\alias{gev_qq}
\alias{gev_rl}
\alias{gev_his}
\alias{pp_and_qq_plot_fn}
\alias{return_level_plot_fn}
\alias{density_plot_fn}
\alias{uniform_confidence_bands}
\alias{gev_confidence_bands}
\alias{merge_two_lists}
\alias{rm_NA_rows}
\alias{cbind_na_for_list_of_matrices}
\alias{make_seasonal_positions}
\alias{get_day}
\alias{get_month}
\alias{get_year}
\title{Internal evmissing functions}
\usage{
negated_gev_loglik(parameters, maxima_notNA, adjust, big_val = Inf)

weighted_negated_gev_loglik(parameters, maxima, weights, big_val = Inf)

discard_maxima(maxima_notNA, discard)

negated_gev_loglik_ret_levs(
  parameters,
  maxima_notNA,
  adjust,
  m,
  npy,
  big_val = Inf
)

weighted_negated_gev_loglik_ret_levs(
  parameters,
  maxima,
  weights,
  m,
  npy,
  big_val = Inf
)

find_block_i_disjoint(i, data, block_length)

find_maxima_notNA_disjoint_pseudo(i, data, block_length)

find_maxima_notNA_disjoint(i, data, block_length)

x_donates_to_y(x, y, full)

x_donates_to_y_seasonal(x, y, full)

find_pseudo_maxima_block_length(data, block_length, full, sliding, seasonal)

x_donates_to_y_block(x, y, full, sliding)

x_donates_to_y_seasonal_block(x, y, full, sliding)

adjust_seasonal_positions(x, y)

find_pseudo_maxima_block(data, block, full, sliding, seasonal)

maxima_block_length(data, block_length, pseudo)

maxima_block(data, block, pseudo)

rhat(parameters, maxima_notNA)

negated_gev_loglik_ts(parameters, maxima_notNA, pseudo, fixed_r, big_val = Inf)

faster_profile_ci(
  negated_loglik_fn,
  which = 1,
  level,
  mle,
  ci_sym_mat,
  inc,
  epsilon,
  ci_init,
  ...
)

profile_ci(negated_loglik_fn, which = 1, level, mle, inc, epsilon, ...)

box_cox(x, lambda = 1, gm = 1, lambda_tol = 1e-06)

box_cox_vec(x, lambda = 1, lambda_tol = 1e-06)

box_cox_deriv(x, lambda = 1, lambda_tol = 1/50, poly_order = 3)

gev_init(maxima_notNA, init_method = "quartiles")

gev_profile_init(data, mu, sigma, xi)

call_gev_profile_init(data, which, par_value)

return_level_profile_init(gev_object, level, mult, epsilon, m, npy)

gev_pp(x, adjust, level, ...)

gev_qq(x, adjust, level, ...)

gev_rl(x, adjust, m, level, profile, num, npy, ...)

gev_his(x, adjust, ...)

pp_and_qq_plot_fn(x, y, bands, which, ..., xlab, ylab = "empirical", main)

return_level_plot_fn(
  x,
  y,
  empirical,
  at,
  labels,
  ...,
  type = "l",
  xlab = "return period",
  ylab = "return level",
  main = "return level plot",
  xaxt = "n"
)

density_plot_fn(
  x,
  mle,
  mu_full,
  sigma_full,
  ...,
  xlab = "block maxima",
  ylab = "density",
  main = "density plot"
)

uniform_confidence_bands(b, k, level)

gev_confidence_bands(b, k, level, mle)

merge_two_lists(list1, list2)

rm_NA_rows(x)

cbind_na_for_list_of_matrices(mat_list)

make_seasonal_positions(block, block_length)

get_day(x)

get_month(x)

get_year(x)
}
\description{
Internal evmissing functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
