% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev_ts.R
\name{gev_ts}
\alias{gev_ts}
\title{GEV ML Inference with Adjustment for Missing Data (Stationary Sequences)}
\usage{
gev_ts(
  data,
  block_length,
  block,
  pseudo = TRUE,
  full = FALSE,
  sliding = TRUE,
  seasonal = sliding,
  init = "quartiles",
  ...
)
}
\arguments{
\item{data}{Either
\itemize{
\item a numeric vector containing a time series of raw data,
\item an object returned from a call to \code{\link{block_maxima}} with \code{pseudo = TRUE},
such as \code{\link{LaPlagnePrecipMaximaSeasonal}}.
}}

\item{block_length}{A numeric scalar. Used calculate the maxima of
disjoint blocks of \code{block_length} contiguous values in the vector \code{data}.
If \code{length(data)} is not an integer multiple of \code{block_length} then
the values at the end of \code{data} that do not constitute a complete block
of length \code{block_length} are discarded, without warning.}

\item{block}{A numeric vector with the same length as \code{data}. The value of
\code{block[i]} indicates the block into which \code{data[i]} falls. The block
lengths implied by \code{block} may differ by at most 1. For example,
\code{block[i]} could give the block in which observation \code{i} was observed,
with block lengths of 366 for leap years and 365 for other years.}

\item{pseudo}{A logical scalar. If \code{pseudo = TRUE} then the pseudo-maxima
returned from \code{\link{block_maxima}} are used to estimate the value of
\eqn{r_i} for an incomplete, partially-observed block. See \strong{Details}
and the \strong{Details} section of \code{\link{block_maxima}}. If \code{pseudo = FALSE} then
the ratio \eqn{n_i/n} is used, as in \code{\link[=gev_mle]{gev_mle()}}.}

\item{full, sliding, seasonal}{Arguments passed to \code{\link{block_maxima}} to
determine how the pseudo-maxima are calculated.}

\item{init}{Either a character scalar, one of \code{"quartiles"} or \code{"moments"},
or a numeric vector of length 3 giving initial estimates of the GEV
location, scale and shape parameters: \eqn{\mu}, \eqn{\sigma} and
\eqn{\xi}. If \code{init = "quartiles"} then initial estimates of \eqn{\mu} and
\eqn{\sigma} are based on sample quartiles of block maxima, ignoring the
underlying numbers of non-missing raw data, and a value of 0 for
\eqn{\xi}. If \code{init = "moments"} then instead we use the sample mean and
variance of these maxima and an initial value of 0.1 for \eqn{\xi}.}

\item{...}{Further arguments to be passed to \code{\link[stats:optim]{stats::optim}}.}
}
\value{
A list, returned from \code{\link[stats:optim]{stats::optim}} (the MLEs are in the
component \code{par}), with the additional components:
\itemize{
\item \code{loglik}: value of the maximised log-likelihood.
\item \code{vcov}: estimated variance-covariance matrix of the parameters.
\item \code{se}: estimated standard errors of the parameters.
\item \code{maxima}: the vector of block maxima used to fit the model.
\item \code{notNA}: the number of non-missing raw values on which the maxima in
\code{maxima} are based.
\item \code{n}: the maximal block length, that is, the largest number of values that
could have been observed in each of these blocks.
\item \code{rvec}: a vector of the values used for \eqn{r_1, ..., r_b}, where \eqn{b}
is the number of blocks. The content depends on the argument \code{pseudo}.
\item \code{rhats}: if \code{pseudo = TRUE}, a vector of the subset of \code{rvec} for
partially-observed blocks only. The attributes \code{"propn_notNA"} and
\code{"unconstrained"} give, respectively, the values of \eqn{n_i/n} for these
blocks and the estimates of \eqn{r_i} before they are constrained to lie
in the interval \eqn{[n_i/n, 1]}.
\item \code{sliding}: the input argument \code{sliding}.
}

The call to \code{gev_ts} is provided in the attribute \code{"call"}.

The class of the returned object is \code{c("evmissing", "gev_ts", "list")}.
}
\description{
Fits a GEV distribution to block maxima using maximum likelihood estimation,
making an adjustment for the locations of missing raw values in each block.
The GEV location and scale parameters are adjusted to reflect the positions
of missing raw values and time series dependence in the data.
}
\details{
If \code{data} is a numeric vector then exactly one of the arguments
\code{block_length} or \code{block} must be supplied. The parameters are fitted
using maximum likelihood estimation.

The adjustment for the numbers of non-missing values underlying the block
maxima is based on the strong assumption that missing values occur
completely at random. We suppose that a block maximum \eqn{M_n} based on
a full (no missing raw values) block of length \eqn{n} has a
\eqn{\text{GEV}(\mu, \sigma, \xi)} distribution, with distribution function
\eqn{G(x)}. Let \eqn{n_i} be the number of non-missing values in block \eqn{i}
and let \eqn{M_{n_i}} denote the block maximum of such a block. We suppose
that, \eqn{M_{n_i}} has a \eqn{\text{GEV}(\mu(r_i), \sigma(r_i), \xi)}
distribution, where \deqn{\mu(r_i) = \mu + \sigma [r_i^\xi -1] / \xi,}
\deqn{\sigma(r_i) = \sigma r_i^\xi,} for some \eqn{n_i/n \leq r_i \leq 1}.
These expressions are based on the \eqn{M_{n_i}} having approximately a
GEV distribution with distribution function \eqn{G(x)^{r_i}}.

For a full block, \eqn{r_i = 1}. If \code{pseudo = TRUE}, then, for a
partially-observed block, the value of \eqn{r_i} is estimated using the
pseudo-maxima returned from \code{\link{block_maxima}} and the GEV distribution
function based on the current value of \eqn{(\mu, \sigma, \xi)} in the
optimisation routine. Suppose that we have a vector \eqn{M_i} of
pseudo-maxima resulting from a particular partially-observed block \eqn{i}.
It can be shown that the components of \eqn{V_i = -\log G(M_i)} each have an
exponential distribution with mean \eqn{1/r_i}. We estimate \eqn{r_i} using
the reciprocal of the mean of the values in \eqn{V_i}.

The negated log-likelihood is minimised using a call to
\code{\link[stats:optim]{stats::optim}} with \code{hessian = TRUE}. If \code{\link[stats:optim]{stats::optim}} throws an error
then a warning is produced and the returned object has \code{NA} values for
the components \code{par}, \code{loglik}, \code{vcov} and \code{se} and an extra component
\code{optim_error} containing the error message. If the estimated observed
information matrix is singular then a warning is produced and the returned
object has \code{NA} values for the components \code{vcov} and \code{se}.
}
\examples{
set.seed(1632026)
blocks <- 50
block_length <- 90
missing_args <- list(p0miss = 0.5, min = 0, max = 0.4)
sdata <- sim_data(blocks = blocks, block_length = block_length,
                  missing_args = missing_args)

pf <- gev_ts(sdata$data_miss, block_length = 90, pseudo = FALSE)
pt <- gev_ts(sdata$data_miss, block_length = 90, pseudo = TRUE)
pt2 <- gev_ts(sdata$data_miss, block_length = 90, pseudo = TRUE,
              seasonal = FALSE)
ptb <- gev_ts(sdata$data_miss, block = sdata$block, pseudo = TRUE)
ptb2 <- gev_ts(sdata$data_miss, block = sdata$block, pseudo = TRUE,
               seasonal = FALSE)

## La Plagne data

# Precipitation
p <- gev_ts(LaPlagnePrecipMaximaSeasonal)
coef(p)

# Snow depth
s <- gev_ts(LaPlagnePrecipMaximaSeasonal)
coef(s)
}
\seealso{
\code{\link{gev_mle}} provides an adjustment for missing data in the
case where the raw data can be assumed to be independent and identically
distributed.
}
