% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LaPlagnePrecipMaximaSeasonal}
\alias{LaPlagnePrecipMaximaSeasonal}
\title{Winter Precipitation Maxima and Seasonal Missing Information at La Plagne, France}
\format{
\code{LaPlagnePrecipMaximaSeasonal} is a list. See \code{\link{block_maxima}}.
}
\usage{
data(LaPlagnePrecipMaximaSeasonal)
}
\description{
These data are the result of a call
\code{block_maxima(data = LaPlagne$rr24, block = LaPlagne$winter, pseudo = TRUE, seasonal = TRUE)}
to \code{\link{block_maxima}}.
}
\seealso{
\code{\link{LaPlagne}}.
}
\keyword{datasets}
