% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LaPlagneSnowMissings}
\alias{LaPlagneSnowMissings}
\title{Winter Snow Missingness at La Plagne, France}
\format{
\code{LaPlagneSnowMissings} is a data frame with 517 rows and the 2
variables:
\itemize{
\item \code{winter}: the winter in which a missing snow depth value occurs.
\item \verb{day of winter}: the day of the winter on which the missing value occurs.
}
}
\usage{
data(LaPlagneSnowMissings)
}
\description{
Provides the days on which the snow depth variable is missing,
that is, the winter in which is occurs and the day of this winter, where
day 1 is 1 Decemeber, day 2 is 2 December etc.
}
\examples{
# Pattern of missingness over and within winters
plot(LaPlagneSnowMissings, pch = "_", main = "missing snow depth")
}
\seealso{
\code{\link{LaPlagne}}.
}
\keyword{datasets}
